feat: add DigitalVisaFeature and refactor config package structure
