+++ service/src/main/java/com/aterrizar/service/checkin/feature/DigitalVisaFeature.java
@@ -1,0 +1,5 @@
+package com.aterrizar.service.checkin.feature;
+
+public interface DigitalVisaFeature {
+    boolean isCountryAvailable(String countryCode);
+}
